% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{AphisRumicisDerrisMalaccensis}
\alias{AphisRumicisDerrisMalaccensis}
\title{data on the toxicity to Aphis rumicis of an ether extract of Derris malaccensis}
\usage{
AphisRumicisDerrisMalaccensis
}
\description{
data on the toxicity to Aphis rumicis of an ether extract of Derris malaccensis
}
\details{
\itemize{
  \item concentration. concentration
  \item n. number of insects
  \item r. number of observed affected
}
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve. pp 238.
Cambridge University Press

Martin, J. T ., 1940
The problem of the evaluation of rotenone-containing plants. V. The relative
toxicities of different species of derris. Ann. Appl. Biol. 27, 274-94.
}
\keyword{data}

