% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{CalculateLCn}
\alias{CalculateLCn}
\title{Calculate LC for N between 0 (LC0) and 100 (LC100)}
\usage{
CalculateLCn(x, n, r, N = 50)
}
\arguments{
\item{x}{numeric, log concentration}

\item{n}{numeric, number of insects}

\item{r}{numeric, number of observed affected}

\item{N}{numeric, Lethal Concentration "N"}
}
\value{
the LC for n between 0 and 100
}
\description{
Returns the LC for n between 0 and 100
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

