% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{Dunnett.t.Statistic}
\alias{Dunnett.t.Statistic}
\title{Critical Values of Dunnett's t Statistic}
\usage{
Dunnett.t.Statistic
}
\description{
Critical Values of Dunnett's t Statistic, Two-Tailed Comparisons
}
\details{
Critical Values of Dunnett's t Statistic - data columns
\itemize{
  \item df. Degress of freedom.
  \item alpha. significance level.
  \item 2. k=2, Number of Treatment Means,  Including Control.
  \item 3. k=3, Number of Treatment Means,  Including Control.
  \item 4. k=4, Number of Treatment Means,  Including Control.
  \item 5. k=5, Number of Treatment Means,  Including Control.
  \item 6. k=6, Number of Treatment Means,  Including Control.
  \item 7. k=7, Number of Treatment Means,  Including Control.
  \item 8. k=8, Number of Treatment Means,  Including Control.
  \item 9. k=9, Number of Treatment Means,  Including Control.
  \item 10. k=10, Number of Treatment Means,  Including Control.
}
}
\author{
Jose Gama
}
\references{
C. W. Dunnett, 1964.
New tables for multiple comparisons with a control.
Biometrics 20. 482–491.
}
\keyword{data}

