% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSK.R
\name{erfinv}
\alias{erfinv}
\title{Inverse error function}
\usage{
erfinv(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
the inverse error function
}
\description{
Returns the inverse error function
}
\examples{
erfinv(1:10)
}
\author{
Jose Gama
}
\references{
Abramowitz and Stegun 29.2.29
\url{http://stat.ethz.ch/R-manual/R-devel/library/stats/html/Normal.html}
}

