% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableIFinney1964}
\alias{GenTableIFinney1964}
\title{Generate table I from Finney1964 "Transformation of percentages to probits"}
\usage{
GenTableIFinney1964()
}
\value{
table I from Finney1964 "Transformation of percentages to probits"
\itemize{
  \item Percentage. Percentage.
  \item Col0.0. Column for 0.0
  \item Col0.1. Column for 0.1
  \item Col0.2. Column for 0.2
  \item Col0.3. Column for 0.3
  \item Col0.4. Column for 0.4
  \item Col0.5. Column for 0.5
  \item Col0.6. Column for 0.6
  \item Col0.7. Column for 0.7
  \item Col0.8. Column for 0.8
  \item Col0.9. Column for 0.9
}
}
\description{
Generates table I from Finney1964 "Transformation of percentages to probits"
}
\examples{
GenTableIFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

