% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableIIFinney1964}
\alias{GenTableIIFinney1964}
\title{Generate table II from Finney1964 "The weighting coefficient and Q/Z"}
\usage{
GenTableIIFinney1964()
}
\value{
table II from Finney1964 "The weighting coefficient and Q/Z"
\itemize{
  \item Y. expected probit
  \item Q/Z.
  \item C=0. 0% of natural mortality
  \item C=1. 1% of natural mortality
...
  \item C=89. 89% of natural mortality
  \item C=90. 90% of natural mortality
}
}
\description{
Generates table II from Finney1964 "The weighting coefficient and Q/Z"
}
\examples{
GenTableIIFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

