% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableIIIFinney1964}
\alias{GenTableIIIFinney1964}
\title{Generate table III from Finney1964 "Maximum and minimum working probits and range"}
\usage{
GenTableIIIFinney1964()
}
\value{
table III from Finney1964 "Maximum and minimum working probits and range"
\itemize{
  \item Ymin. Minimum working probit - expected
  \item Y0. Minimum working probit - Y0 = Y-P/Z
  \item Yrange. Range 1/Z
  \item Y100. Maximum working probit - Y100 = Y+Q/Z
  \item Ymax. Maximum working probit - expected
}
}
\description{
Generates table III from Finney1964 "Maximum and minimum working probits and range"
}
\examples{
GenTableIIIFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

