% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableIVFinney1964}
\alias{GenTableIVFinney1964}
\title{Generate table IV from Finney1964 "Working probits"}
\usage{
GenTableIVFinney1964()
}
\value{
table IV from Finney1964 "Working probits"
\itemize{
  \item Kill%. % kill
  \item Col2 Expected probit 2.0
  \item Col2.1 Expected probit 2.1
...
  \item Col7.8 Expected probit 7.8
  \item Col7.9 Expected probit 7.9
}
}
\description{
Generates table IV from Finney1964 "Working probits"
}
\examples{
GenTableIVFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

