% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableIXFinney1964}
\alias{GenTableIXFinney1964}
\title{Generate table IX from Finney1964 "Minimum Working Probit, Range, and Weighting Coefficient for Inverse Sampling"}
\usage{
GenTableIXFinney1964()
}
\value{
table IX from Finney1964 "Minimum Working Probit, Range, and Weighting Coefficient for Inverse Sampling"
\itemize{
  \item Y. Expected probit
  \item MinWorkProbit. Minimum working probit
  \item Range. Range 1/Z
  \item WeightingCoef. Weighting Coefficient
}
}
\description{
Generates table IX from Finney1964 "Minimum Working Probit, Range, and Weighting Coefficient for Inverse Sampling"
}
\examples{
GenTableIXFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

