% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableVFinney1964}
\alias{GenTableVFinney1964}
\title{Generate table V from Finney1964 "The Probability, P, the Ordinate, Z, and Z^2"}
\usage{
GenTableVFinney1964()
}
\value{
table V from Finney1964 "The Probability, P, the Ordinate, Z, and Z^2"
\itemize{
  \item Y. Expected probit
  \item P. Probability P of expected probit
  \item Z. Ordinate to the normal distribution corresponding to the probability P
  \item Z^2. Z^2
}
}
\description{
Generates table V from Finney1964 "The Probability, P, the Ordinate, Z, and Z^2"
}
\examples{
GenTableVFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

