% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableVIFinney1964}
\alias{GenTableVIFinney1964}
\title{Generate table VI from Finney1964 "Distribution of chi^2"}
\usage{
GenTableVIFinney1964()
}
\value{
table VI from Finney1964 "Distribution of chi^2"
\itemize{
  \item Deg.freedom. Degrees of freedom
  \item 0.9. Probability 0.9
  \item 0.7. Probability 0.7
  \item 0.5. Probability 0.5
  \item 0.3. Probability 0.3
  \item 0.1. Probability 0.1
  \item 0.05. Probability 0.05
  \item 0.02. Probability 0.02
  \item 0.01. Probability 0.01
  \item 0.001. Probability 0.001
}
}
\description{
Generates table VI from Finney1964 "Distribution of chi^2"
}
\examples{
GenTableVIFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

