% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableVIIFinney1964}
\alias{GenTableVIIFinney1964}
\title{Generate table VII from Finney1964 "Distribution of t"}
\usage{
GenTableVIIFinney1964()
}
\value{
table VII from Finney1964 "Distribution of t"
\itemize{
  \item Deg.freedom. Degrees of freedom
  \item 0.9. Probability 0.9
  \item 0.7. Probability 0.7
  \item 0.5. Probability 0.5
  \item 0.3. Probability 0.3
  \item 0.1. Probability 0.1
  \item 0.05. Probability 0.05
  \item 0.02. Probability 0.02
  \item 0.01. Probability 0.01
  \item 0.001. Probability 0.001
}
}
\description{
Generates table VII from Finney1964 "Distribution of t"
}
\examples{
GenTableVIIFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

