% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{GenTableVIIIFinney1964}
\alias{GenTableVIIIFinney1964}
\title{Generate table VIII from Finney1964 "The Weighting Coefficient in Wadley's Problem"}
\usage{
GenTableVIIIFinney1964()
}
\value{
table VIII from Finney1964 "The Weighting Coefficient in Wadley's Problem"
\itemize{
  \item Y. Expected probit
  \item w. Weighting Coefficient
}
}
\description{
Generates table VIII from Finney1964 "The Weighting Coefficient in Wadley's Problem"
}
\examples{
GenTableVIIIFinney1964()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

