% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{IsMonotonicallyDecreasing}
\alias{IsMonotonicallyDecreasing}
\title{Determine if a series is monotonically decreasing}
\usage{
IsMonotonicallyDecreasing(p)
}
\arguments{
\item{p}{numeric vector}
}
\value{
True is the series is monotonically decreasing
}
\description{
Returns TRUE if all proportions are in a monotonically decreasing sequence
}
\examples{
IsMonotonicallyDecreasing(1:10)
IsMonotonicallyDecreasing(6:2)
IsMonotonicallyDecreasing(c(1,3,2))
}
\author{
Jose Gama
}
\references{
Hamilton, m.a., R.c. Russo, and r.v. Thurston, 1977.
Trimmed spearman-karber method for estimating median
Lethal concentrations in toxicity bioassays.
Environ. Sci. Technol. 11(7): 714-719;
Correction 12(4):417 (1978).
}

