% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{MakeMonotonicallyDecreasing}
\alias{MakeMonotonicallyDecreasing}
\title{Make monotonically decreasing sequence}
\usage{
MakeMonotonicallyDecreasing(matrixExpoResp)
}
\arguments{
\item{matrixExpoResp}{numeric vector or matrix}
}
\value{
monotonically decreasing sequence
}
\description{
Returns a monotonically decreasing sequence
}
\author{
Jose Gama
}
\references{
Hamilton, m.a., R.c. Russo, and r.v. Thurston, 1977.
Trimmed spearman-karber method for estimating median
Lethal concentrations in toxicity bioassays.
Environ. Sci. Technol. 11(7): 714-719;
Correction 12(4):417 (1978).
}

