% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{MakeMonotonicallyIncreasing}
\alias{MakeMonotonicallyIncreasing}
\title{Smoothed Mortality Proportion (monotonically increasing sequence)}
\usage{
MakeMonotonicallyIncreasing(matrixExpoResp)
}
\arguments{
\item{matrixExpoResp}{numeric vector or matrix}
}
\value{
The Smoothed Mortality Proportion (monotonically increasing sequence)
}
\description{
Returns the Smoothed Mortality Proportion (monotonically increasing sequence)
}
\author{
Jose Gama
}
\references{
Hamilton, m.a., R.c. Russo, and r.v. Thurston, 1977.
Trimmed spearman-karber method for estimating median
Lethal concentrations in toxicity bioassays.
Environ. Sci. Technol. 11(7): 714-719;
Correction 12(4):417 (1978).
}

