% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitApproxStandardErrorOfDosage}
\alias{ProbitApproxStandardErrorOfDosage}
\title{Approximate Standard Error of dosage}
\usage{
ProbitApproxStandardErrorOfDosage(b, Snw)
}
\arguments{
\item{b}{numeric, rate of increase of probit value per unit increase in x}

\item{Snw}{numeric, sum of nw}
}
\value{
Approximate Standard Error of dosage
}
\description{
Approximate Standard Error of dosage
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

