% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitChi}
\alias{ProbitChi}
\title{Estimate the column for Chi calculation}
\usage{
ProbitChi(r, n, P)
}
\arguments{
\item{r}{numeric vector, number of observed affected}

\item{n}{numeric vector, number of insects}

\item{P}{numeric vector, Probability P of expected probit}
}
\value{
numeric vector
}
\description{
Estimates the column for Chi calculation
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

