% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitEPA}
\alias{ProbitEPA}
\title{Probit estimation similar to the EPA's Ecological Exposure Research Division (EERD) tool}
\usage{
ProbitEPA(toxData, retData = FALSE, showOutput = TRUE)
}
\arguments{
\item{toxData}{numeric matrix, matrix with concentration, n ,r columns}

\item{retData}{logic, return the results in a list}

\item{showOutput}{logic, show results in the console}
}
\value{
Probit estimation regression
}
\description{
Probit estimation similar to the EPA's Ecological Exposure Research Division (EERD) tool
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

