% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitFiducialLimits}
\alias{ProbitFiducialLimits}
\title{Probit Fiducial Limits}
\usage{
ProbitFiducialLimits(Vm, m, tPercent = 5, roundFinney = FALSE)
}
\arguments{
\item{Vm}{numeric, variance of the logarithm}

\item{m}{numeric, logLD50}

\item{tPercent}{numeric, probability level}

\item{roundFinney}{logic, round as in Finney's book}
}
\value{
Probit Fiducial Limits
}
\description{
Probit Fiducial Limits
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

