% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitFinney}
\alias{ProbitFinney}
\title{Probit estimation regression with Finney's method}
\usage{
ProbitFinney(toxData, tPercent = 5, showPlot = FALSE, roundFinney = FALSE)
}
\arguments{
\item{toxData}{numeric matrix, matrix with concentration, n ,r columns}

\item{tPercent}{numeric, probability level}

\item{showPlot}{logic, show regression line - plot}

\item{roundFinney}{logic, round as in Finney's book}
}
\value{
Probit estimation regression
}
\description{
Probit estimation regression with Finney's method
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

