% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitStandardErrorOfDosage}
\alias{ProbitStandardErrorOfDosage}
\title{Standard Error of dosage}
\usage{
ProbitStandardErrorOfDosage(varianceDosage)
}
\arguments{
\item{varianceDosage}{numeric, Variance of dosage}
}
\value{
Standard Error of dosage
}
\description{
Standard Error of dosage
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

