% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitStandardErrorRate}
\alias{ProbitStandardErrorRate}
\title{Standard Error of rate of increase of probit value per unit increase in x}
\usage{
ProbitStandardErrorRate(n, w, x, xbar)
}
\arguments{
\item{n}{numeric, number of insects}

\item{w}{numeric, weighting coefficients}

\item{x}{numeric, log concentration}

\item{xbar}{numeric, mean dosage}
}
\value{
Standard Error of rate of increase of probit value per unit increase in x
}
\description{
Standard Error of rate of increase of probit value per unit increase in x
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

