% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitVALUEg}
\alias{ProbitVALUEg}
\title{Probit value "g"}
\usage{
ProbitVALUEg(b, n, w, x, xbar, tPercent)
}
\arguments{
\item{b}{numeric, rate of increase of probit value per unit increase in x}

\item{n}{numeric, number of insects}

\item{w}{numeric, weighting coefficients}

\item{x}{numeric, log concentration}

\item{xbar}{numeric, mean dosage}

\item{tPercent}{numeric, probability level}
}
\value{
Probit value "g"
}
\description{
Probit value "g"
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

