% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitVarianceDosage}
\alias{ProbitVarianceDosage}
\title{Variance of dosage}
\usage{
ProbitVarianceDosage(b, m, n, w, x, xbar)
}
\arguments{
\item{b}{numeric, rate of increase of probit value per unit increase in x}

\item{m}{numeric, dosage}

\item{n}{numeric, number of insects}

\item{w}{numeric, weighting coefficients}

\item{x}{numeric, log concentration}

\item{xbar}{numeric, mean dosage}
}
\value{
Variance of dosage
}
\description{
Variance of dosage
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

