% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ProbitZ4dec}
\alias{ProbitZ4dec}
\title{Calculate the ordinate to the normal distribution corresponding to the probability P, exactly like Finney's}
\usage{
ProbitZ4dec(Y)
}
\arguments{
\item{Y}{numeric, expected probit}
}
\value{
the ordinate to the normal distribution corresponding to the probability P
}
\description{
Returns the ordinate to the normal distribution corresponding to the probability P
with the exact same results as Finney's
}
\examples{
# expected probit Y = 6.2
Y <- 6.2
ProbitZ4dec(Y)
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press. Formula 3.5.
}

