% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ScaleArcher}
\alias{ScaleArcher}
\title{Archer Scale for assessment of leaf damage}
\usage{
ScaleArcher(percentAffected)
}
\arguments{
\item{percentAffected}{numeric vector}
}
\value{
Archer Scale for assessment of leaf damage
}
\description{
Archer Scale for assessment of leaf damage
}
\author{
Jose Gama
}
\references{
Archer, T.L., 1987
Techniques for screening maize for resistance to mites. pp.178-183. In:
Mihn, J.A., Wiseman, B.R. and Davis, F.M. (Eds.). Proceedings of the International
symposium on methodologies for developing host plant resistance to maize insects.
CIMMYT, Mexico.
}

