% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ScaleGauhlStover}
\alias{ScaleGauhlStover}
\title{Gauhl’s modification of Stover’s severity scoring system}
\usage{
ScaleGauhlStover(percentShowingSymptoms)
}
\arguments{
\item{percentShowingSymptoms}{numeric, proportion of the leaf area showing symptoms}
}
\value{
Gauhl-Stover scale
}
\description{
Gauhl’s modification of Stover’s severity scoring system
}
\author{
Jose Gama
}
\references{
Gauhl F., 1994
Epidemiology and ecology of black Sigatoka (Mycosphaerella fijiensis Morlet)
on plantain and banana (Musa spp.) in Costa Rica, Central America.
INIBAP, Montpellier, France. 120pp).
}

