% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{ScaleHorsfallBarratt}
\alias{ScaleHorsfallBarratt}
\title{Horsfall-Barratt Scale for Measuring Plant Disease}
\usage{
ScaleHorsfallBarratt(percentAffected)
}
\arguments{
\item{percentAffected}{numeric vector}
}
\value{
Horsfall-Barratt Scale for Measuring Plant Disease
}
\description{
Horsfall-Barratt Scale for Measuring Plant Disease
}
\author{
Jose Gama
}
\references{
Horsfall, J. G.; Barratt, R. W., 1945
An Improved Grading System for Measuring Plant Disease. Phytopathology.
}

