% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{SheepsheadMinnow40SK}
\alias{SheepsheadMinnow40SK}
\title{Mortality data from a fathead minnow larval survival and growth test (40 organisms per concentration)}
\usage{
SheepsheadMinnow40SK
}
\description{
Mortality data from a fathead minnow larval survival and growth test (40 organisms per concentration)
}
\details{
Mortality data from a fathead minnow larval survival and growth test - data columns
\itemize{
  \item Concentration. Concentration.
  \item Mortality. Mortality
}
}
\author{
Jose Gama
}
\references{
USEPA, 2002
Short-term Methods for Estimating the Chronic Toxicity of Effluents and Receiving Waters to Freshwater Organisms.
4th Edition,USEPA,Office of Water,October 2002,EPA 821-R-02-013
TABLE J1. pp 312
}
\keyword{data}

