% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{SpearmanKarber}
\alias{SpearmanKarber}
\title{Spearman Karber estimation}
\usage{
SpearmanKarber(toxData, N, retData = FALSE, showOutput = TRUE,
  showPlot = TRUE)
}
\arguments{
\item{toxData}{numeric matrix, matrix with concentration, n ,r columns}

\item{N}{numeric, number of organisms}

\item{retData}{logic, return the results in a list}

\item{showOutput}{logic, show results in the console}

\item{showPlot}{logic, show regression line - plot}
}
\value{
Spearman Karber estimation
}
\description{
Spearman Karber estimation
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

