% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{Table1Finney1964}
\alias{Table1Finney1964}
\title{Transformation of Percentages to Probits, table I of Finney, 1964}
\usage{
Table1Finney1964
}
\description{
Transformation of Percentages to Probits, table I of Finney, 1964
}
\details{
Transformation of Percentages to Probits - data columns
\itemize{
  \item Percentage. Percentage.
  \item Col0.0. Column for 0.0
  \item Col0.1. Column for 0.1
  \item Col0.2. Column for 0.2
  \item Col0.3. Column for 0.3
  \item Col0.4. Column for 0.4
  \item Col0.5. Column for 0.5
  \item Col0.6. Column for 0.6
  \item Col0.7. Column for 0.7
  \item Col0.8. Column for 0.8
  \item Col0.9. Column for 0.9
}
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}
\keyword{data}

