% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{Table2Finney1964}
\alias{Table2Finney1964}
\title{The Weighting Coefficient and Q/Z, table II of Finney, 1964}
\usage{
Table2Finney1964
}
\description{
The Weighting Coefficient and Q/Z, table II of Finney, 1964
}
\details{
The Weighting Coefficient and Q/Z - data columns
\itemize{
  \item Y. expected probit
  \item Q/Z.
  \item C=0. 0% of natural mortality
  \item C=1. 1% of natural mortality
...
  \item C=89. 89% of natural mortality
  \item C=90. 90% of natural mortality
}
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}
\keyword{data}

