% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{Table3Finney1964}
\alias{Table3Finney1964}
\title{Maximum and Minimum working probits and Range, table III of Finney, 1964}
\usage{
Table3Finney1964
}
\description{
Maximum and Minimum working probits and Range, table III of Finney, 1964
}
\details{
Maximum and Minimum working probits and Range - data columns
\itemize{
  \item Ymin. Minimum working probit - expected
  \item Y0. Minimum working probit - Y0 = Y-P/Z
  \item Yrange. Range 1/Z
  \item Y100. Maximum working probit - Y100 = Y+Q/Z
  \item Ymax. Maximum working probit - expected
}
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}
\keyword{data}

