% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{Table4Finney1964}
\alias{Table4Finney1964}
\title{Working probits, table IV of Finney, 1964}
\usage{
Table4Finney1964
}
\description{
Working probits, table IV of Finney, 1964
}
\details{
Working probits - data columns
\itemize{
  \item Kill%. % kill
  \item Col2 Expected probit 2.0
  \item Col2.1 Expected probit 2.1
...
  \item Col7.8 Expected probit 7.8
  \item Col7.9 Expected probit 7.9
}
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}
\keyword{data}

