% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{Table5Finney1964}
\alias{Table5Finney1964}
\title{The Probability, P, the Ordinate, Z, and Z^2, table V of Finney, 1964}
\usage{
Table5Finney1964
}
\description{
Probability, P, the Ordinate, Z, and Z^2, table V of Finney, 1964
}
\details{
The Probability, P, the Ordinate, Z, and Z^2 - data columns
\itemize{
  \item Y. Expected probit
  \item P. Probability P of expected probit
  \item Z. Ordinate to the normal distribution corresponding to the probability P
  \item Z^2. Z^2
}
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}
\keyword{data}

