% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{Table8Finney1964}
\alias{Table8Finney1964}
\title{The Weighting Coefficient in Wadley's Problem, table VIII of Finney, 1964}
\usage{
Table8Finney1964
}
\description{
The Weighting Coefficient in Wadley's Problem, table VIII of Finney, 1964
}
\details{
The Weighting Coefficient in Wadley's Problem - data columns
\itemize{
  \item Y. Expected probit
  \item w. Weighting Coefficient
}
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}
\keyword{data}

