% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\docType{data}
\name{Table9Finney1964}
\alias{Table9Finney1964}
\title{Minimum Working Probit, Range, and Weighting Coefficient for Inverse Sampling, table IX of Finney, 1964}
\usage{
Table9Finney1964
}
\description{
Minimum Working Probit, Range, and Weighting Coefficient for Inverse Sampling, table IX of Finney, 1964
}
\details{
Minimum Working Probit, Range, and Weighting Coefficient for Inverse Sampling - data columns
\itemize{
  \item Y. Expected probit
  \item MinWorkProbit. Minimum working probit
  \item Range. Range 1/Z
  \item WeightingCoef. Weighting Coefficient
}
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}
\keyword{data}

