% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{TestMix2poisons}
\alias{TestMix2poisons}
\title{Generate table 26 from Finney1964 "The Function for Planning Tests of Mixtures of Two Poisons"}
\usage{
TestMix2poisons()
}
\value{
table 26 from Finney1964 "The Function for Planning Tests of Mixtures of Two Poisons"
\itemize{
  \item rho. toxicity
  \item 0.1. distance 0.1 log rho in the left of the probit regression line
...
  \item 0.9. distance 0.9 log rho in the left of the probit regression line
}
}
\description{
Generates table 26 from Finney1964 "The Function for Planning Tests of Mixtures of Two Poisons"
}
\examples{
TestMix2poisons()
}
\author{
Jose Gama
}
\references{
Finney D. J., 1964
Probit analysis: a statistical treatment of the sigmoid response curve.
Cambridge University Press
}

