% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecotoxicology.R
\name{WAAPPpestCount}
\alias{WAAPPpestCount}
\title{WAAPP Pest Count scoring system}
\usage{
WAAPPpestCount(percentLeafDamage)
}
\arguments{
\item{percentLeafDamage}{numeric, percentage of leaf damage}
}
\value{
WAAPP Pest Count Score
}
\description{
WAAPP Pest Count scoring system
}
\author{
Jose Gama
}
\references{
Environmental Protection Agency Chemicals Control And Managemenet Centre (ACCRA),  2012
Protocols for the biological evaluation of pesticides on Selected crops grown
in both the humid and sahel regions of West africa.
West Africa Agriculture Productivity Programme (WAAPP).
}

